actors support added
## Actors Support

If you have announceActors enabled and have this running on the other Characters on the same PC, then your DPS History window will show a Party tab that contains the Current Battle DPS for each character or the Last Battle DPS. You can toggle the on and off in the settings. The Party Tab has an option to sort or not.

If you are sorting then it sorts Current Battle On top and by DPS highest DPS on top. The sorting happens at a 30fps interval to allow you to actually read the data. This is to prevent the rows bouncing to fast. You do not have this restriction when not sorting the data is refreshed as you receive it.